docs: address Pi Agent code-review feedback (shared helpers, --approve, deps)
Addresses cubesandbox-bot review on #701:

- Extract run_command / ensure_success / sandbox_identifier / stream_writer
  into a shared _pi_common.py so run_pi_agent / resume_pi_agent /
  network_policy no longer import from one another (env_utils.py stays
  focused on .env/config).
- env_utils.py: pi_command() gains an `approve` flag (default True) so callers
  can disable --approve (which trusts project-local AGENTS.md/CLAUDE.md) in
  high-security workflows.
- requirements.txt: drop unused `rich`; README file trees add _pi_common.py.
- docs: caveat that the direct (`envs=`) key may be cached by Pi under
  /root/.pi/agent and survive pause/resume; prefer the vault flavor.
