+++ examples/pi-agent-integration/requirements.txt
@@ -1,4 +1,3 @@
 e2b>=2.4.1
 cubesandbox>=0.3.0
 python-dotenv>=1.0.0
-rich>=13.0
